+++ tests/union_in_struct.c
@@ -3,42 +3,6 @@
 #include <stdio.h>
 #include "tests.h"
 
-struct SHA3 {
-  union {
-    double iY;
-    double dY;
-  } uY;
-  float ffY;
-};
-
-union unknown {
-  double i2;
-  double d2;
-};
-struct SHA32 {
-  union unknown u2;
-  float ff2;
-};
-
-
-void union_inside_struct()
-{
-	diag("Union inside struct");
-	struct SHA3 sha;
-	sha.ffY  = 12.444;
-	sha.uY.iY = 4;
-	is_eq(sha.uY.iY, 4);
-	is_eq(sha.uY.dY, 4);
-	is_eq(sha.ffY , 12.444);
-
-	struct SHA32 sha2;
-	sha2.ff2  = 12.444;
-	sha2.u2.i2 = 4;
-	is_eq(sha2.u2.i2, 4);
-	is_eq(sha2.u2.d2, 4);
-	is_eq(sha2.ff2 , 12.444);
-}
-
 typedef struct FuncDestructor FuncDestructor;
 struct FuncDestructor {
 	int i;
